==> 1A_Data_Wrangling.ipynb <==
import pandas as pd

import warnings
def ignore_warn(*args, **kwargs):
    pass
warnings.warn = ignore_warn

# Table of Content
1. [COMPUSTAT](#COMPUSTAT-Data)
2. [CRSP](#CRSP-Data)

**Note:** Kindly run only one of the database

## COMPUSTAT Data

df_compustat = pd.read_csv('data/COMPUSTAT_20220423.csv')
df_compustat

### Data Wrangling - COMPUSTAT

'''
The default web query has boxes checked for Industry Formats (INDFMT) “INDL” and “FS.”
FS: Financial Services (includes banks, insurance companies, broker/dealers, real estate and other financial services)
INDL: Industrial (includes companies reporting manufacturing, retail, construction and other commercial operations other than financial services)

The default setting on the web query pulling both of these is the most notable source for duplicate observations.

Pulling both of these formats for a non-financial services firm for a given GVKEY and DATADATE results in:
- The one GKVEY DATADATE observation with INDFMT = INDL for the actual 10-K numbers.
- The second GVKEY DATADATE observation with INDFMT= FS is essentially the same 10-K converted to a “financial services” format.
Source: https://robsonglasscock.wordpress.com/2018/04/12/gvkey-and-datadate-or-fyear-duplicates-in-compustat/

For the remaining conditions, please see the following link:
Link: http://kaichen.work/?p=387
'''

# Removing duplicate entries based on Industry formats
df_compustat = df_compustat[(df_compustat['indfmt']=='INDL') & (df_compustat['datafmt']=='STD') & (df_compustat['popsrc']=='D') & (df_compustat['consol']=='C') & (df_compustat['scf']<4) | (df_compustat['scf']==7)]
df_compustat

# Extracting/Selecting required columns only
cols_required = ['cusip', 'conm', 'gsector', 'ggroup', 'gind', 'gsubind', 'fyear', 'epspx', 'bkvlps',]

df_compustat = df_compustat[cols_required]
df_compustat

#Renaming the CUSIPs - Taking the first 6 characters

df_compustat['cusip'] = df_compustat.cusip.str[:6]
df_compustat['cusip'] = df_compustat['cusip'].astype(str)
df_compustat

# Analyzing the number of NAs in the database

df_compustat.isna().sum()

# Analyzing the number of NAs after temporary dropping the NAs from the columns. This step is done by incrementally adding the columns to dropna function.

df_compustat.dropna(subset=['cusip', 'fyear', 'gsector', 'ggroup', 'gind', 'gsubind', 'epspx', 'bkvlps']).isna().sum()

#Dropping all the rows with NAs
df_compustat.dropna(subset=['cusip', 'fyear', 'gsector', 'ggroup', 'gind', 'gsubind', 'epspx', 'bkvlps'], inplace=True)
df_compustat

# Converting the fyear to Integer type
df_compustat['fyear'] = df_compustat['fyear'].astype(int)
df_compustat.dtypes

#Lagging the variables
df_compustat['epspx'] = df_compustat.sort_values(by=['cusip','fyear'],ascending=False).groupby(['cusip'])['epspx'].shift(-1)
df_compustat['bkvlps'] = df_compustat.sort_values(by=['cusip','fyear'],ascending=False).groupby(['cusip'])['bkvlps'].shift(-1)
df_compustat.dropna(subset=['epspx', 'bkvlps'], inplace=True)
df_compustat.reset_index(drop=True, inplace=True)
df_compustat

# Exporting the data

df_compustat.to_csv('data/Cleaned_Compustat.csv', index=False)

del(df_compustat)

from IPython.display import display_html
def restartkernel() :
    display_html("<script>Jupyter.notebook.kernel.restart()</script>",raw=True)

## CRSP Data

# Extracting/Selecting required columns only
required_cols = ['DATE', 'CUSIP', 'PERMNO', 'SHRCD', 'PRC', 'SHROUT', 'RET']

# Reading the dataset
df_crsp = pd.read_csv('data/msf_new2.csv', usecols=required_cols).rename(columns={'DATE':'date'})
df_crsp

### Data Wrangling - CRSP
- SHRCD means Share Code
    - First Digit = 1 means 'Ordinary Common Shares'
    - Second Digit = 0 means 'No special status found' & Second Digit = 1 means 'No special status ncessary'

#Filtering for common shares & dropping the column
df_crsp = df_crsp[(df_crsp['SHRCD']==10.0) | (df_crsp['SHRCD']==11.0)]
df_crsp.drop(columns=['SHRCD'], inplace=True)
df_crsp

# Checking the number of NAs in the dataset
df_crsp.isna().sum()

# Convert column "PRC" of a DataFrame to numeric

df_crsp['PRC'] = pd.to_numeric(df_crsp['PRC'])
df_crsp.dtypes

# Dropping all the rows with PRC as NA
df_crsp.dropna(subset=['PRC'], inplace=True)
df_crsp

'''
Q. Why is the stock price in CRSP preceded by a minus sign?
A.
- Sometimes you see negative stock prices in CRSP. This means that there was no closing price available for that period.
- Instead, the bid/ask average was used.
- To distinguish the bid/ask averages from actual closing prices, CRSP puts a leading dash in front of the price when the bid/ask average was used.
- If neither price nor bid/ask average is available, Price or Bid/Ask Average is set to zero.
'''

# Converting the PRC to positive value
df_crsp['PRC'] = df_crsp['PRC'].abs()

# Creating a column named 'fyear' - it will be helpful in merging the COMPUSTAT and CRSP dataframes later on
df_crsp['fyear'] = (df_crsp['date']/10000).astype('int')

#Converting the date to Datetime
df_crsp['date'] = pd.to_datetime(df_crsp['date'], format='%Y%m%d')

# Creating a column named 'mktcap' - to calculate the Market Capitalization of the stock at the point in time
df_crsp['mktcap'] = abs(df_crsp['PRC'])*df_crsp['SHROUT']

'''
The first 6 digits of a CUSIP identify the company, digits 7-8 describe the security and the 9th is check digit.
'''

#Renaming the CUSIPs - Taking the first 6 characters
df_crsp['CUSIP'] = df_crsp.CUSIP.str[:6]

# Renaming the columns for consistency with COMPUSTAT
df_crsp.rename(columns={'CUSIP':'cusip', 'PRC':'prc', 'SHROUT':'shrout', 'RET':'ret', 'PERMNO':'permno'}, inplace=True)

#Rearranging the columns
df_crsp = df_crsp[['date', 'fyear', 'cusip', 'permno', 'prc', 'shrout', 'mktcap', 'ret']]

df_crsp

#Lagging the variables
df_crsp['prc'] = df_crsp.sort_values(by=['cusip','date'],ascending=False).groupby(['cusip'])['prc'].shift(-1)
df_crsp['shrout'] = df_crsp.sort_values(by=['cusip','date'],ascending=False).groupby(['cusip'])['shrout'].shift(-1)
df_crsp['mktcap'] = df_crsp.sort_values(by=['cusip','date'],ascending=False).groupby(['cusip'])['mktcap'].shift(-1)
df_crsp['ret'] = df_crsp.sort_values(by=['cusip','date'],ascending=False).groupby(['cusip'])['ret'].shift(-1)

df_crsp.dropna(subset=['prc', 'shrout', 'mktcap', 'ret'], inplace=True)
df_crsp.reset_index(drop=True, inplace=True)
df_crsp = df_crsp[df_crsp['ret']!='C']
df_crsp['ret'] = df_crsp['ret'].astype('float64')
df_crsp

df_crsp.to_csv('data/Cleaned_Crsp.csv', index=False)

from IPython.display import display_html
def restartkernel() :
    display_html("<script>Jupyter.notebook.kernel.restart()</script>",raw=True)
