==> 2A_Merging_Datasets.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import warnings
def ignore_warn(*args, **kwargs):
    pass
warnings.warn = ignore_warn

## Importing cleaned datasets

df_cleaned_compustat = pd.read_csv('data/Cleaned_Compustat.csv', usecols=['fyear', 'cusip', 'gsector', 'ggroup', 'gind', 'gsubind', 'epspx', 'bkvlps'])
df_cleaned_compustat

df_cleaned_crsp = pd.read_csv('data/Cleaned_Crsp.csv')
df_cleaned_crsp

## Merging the datasets

df_final = df_cleaned_crsp.merge(df_cleaned_compustat, left_on=['cusip','fyear'], right_on=['cusip','fyear'], how='left')
df_final

df_final[df_final['fyear']>2010].to_csv('data/CleanData_Merged.csv', index=False)

del(df_final)

from IPython.display import display_html
def restartkernel() :
    display_html("<script>Jupyter.notebook.kernel.restart()</script>",raw=True)
