==> 3_Cleaning_Merged_Data.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

import warnings
def ignore_warn(*args, **kwargs):
    pass
warnings.warn = ignore_warn

## Importing the Clean Merged Data

df_clean_merged = pd.read_csv('data/CleanData_Merged.csv')
df_clean_merged

# Analyzing NAs in dataset
df_clean_merged.dropna(subset=['cusip','gsector']).isna().sum()

# Dropping the NAs in dataset
df_clean_merged.dropna(subset=['cusip','gsector'], inplace=True)
df_clean_merged

df_clean_merged.to_csv('data/Clean_FinalData.csv', index=False)
